requirements.txt
aiohttp==3.6.2
aiomysql==0.0.20
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
bs4==0.0.1
cffi==1.14.2
chardet==3.0.4
cryptography==3.0
discord==1.0.1
discord.py==1.4.1
idna==2.10
multidict==4.7.6
pycparser==2.20
PyMySQL==0.9.2
python-dotenv==0.14.0
six==1.15.0
soupsieve==2.0.1
yarl==1.5.1
